Fait vivre les grottes d'Aurore : plus de puits et de brumes qui suivent, la Chambre d'Aurore, le Rappel
4-6 puits et 2-3 paires par levee, relevee a 48 blocs, brume et puits sur le
panneau. La Chambre d'Aurore : la roche s'ouvre sur une brume dont la jumelle
se leve contre le filon designe. Le Rappel : a la fin, une brume d'or ramene
au jour qui est encore sous terre.

+++ src/main/java/com/emerald/weather/BattueHunt.java
@@ -232,7 +232,7 @@ public final class BattueHunt {
         if (now % 40 == 0) {
             for (ServerPlayer player : level.players()) {
                 PacketDistributor.sendToPlayer(player, new com.emerald.network.VeinSyncPayload(
-                        List.of(beast.blockPosition().asLong()), com.emerald.network.VeinSyncPayload.KIND_PREY));
+                        List.of(beast.blockPosition().asLong()), (long) com.emerald.network.VeinSyncPayload.KIND_PREY));
             }
         }
         // LA CHARGE DU SANGLIER : accule, il rend un coup, puis il refuit
